+++ README.md
@@ -96,6 +96,7 @@ macho_fat,
 [matroska](doc/formats.md#matroska),
 [mp3](doc/formats.md#mp3),
 mp3_frame,
+mp3_frame_tags,
 [mp4](doc/formats.md#mp4),
 mpeg_asc,
 mpeg_es,
@@ -132,7 +133,6 @@ vpx_ccr,
 [wasm](doc/formats.md#wasm),
 wav,
 webp,
-xing,
 [xml](doc/formats.md#xml),
 yaml,
 [zip](doc/formats.md#zip)
