+++ format/format.go
@@ -91,6 +91,7 @@ const (
 	MATROSKA            = "matroska"
 	MP3                 = "mp3"
 	MP3_FRAME           = "mp3_frame"
+	MP3_FRAME_TAGS      = "mp3_frame_tags"
 	MP4                 = "mp4"
 	MPEG_ASC            = "mpeg_asc"
 	MPEG_ES             = "mpeg_es"
@@ -127,7 +128,6 @@ const (
 	WASM                = "wasm"
 	WAV                 = "wav"
 	WEBP                = "webp"
-	XING                = "xing"
 	XML                 = "xml"
 	YAML                = "yaml"
 	ZIP                 = "zip"
