+++ format/mp3/testdata/mp3_frame_tags_info.fqtest
@@ -1,6 +1,6 @@
-# ffmpeg -y -f lavfi -i sine -ac 2 -t 10ms -f mp3 temp.mp3 && fq temp.mp3 '.frame[0].xing | tobits' > xing
-$ fq -d xing dv xing
-    |00 01 02 03 04 05 06 07 08 09 0a 0b 0c 0d 0e 0f|0123456789abcdef|.{}: xing (xing) 0x0-0x9b.7 (156)
+# ffmpeg -y -f lavfi -i sine -ac 2 -t 10ms -f mp3 temp.mp3 && fq temp.mp3 '.frames[0].tags | tobits' > mp3_frame_tags_info
+$ fq -d mp3_frame_tags dv mp3_frame_tags_info
+    |00 01 02 03 04 05 06 07 08 09 0a 0b 0c 0d 0e 0f|0123456789abcdef|.{}: mp3_frame_tags_info (mp3_frame_tags) 0x0-0x9b.7 (156)
 0x00|49 6e 66 6f                                    |Info            |  header: "Info" 0x0-0x3.7 (4)
     |                                               |                |  present_flags{}: 0x4-0x7.7 (4)
 0x00|            00 00 00 0f                        |    ....        |    unused: 0 0x4-0x7.3 (3.4)
@@ -112,23 +112,22 @@ $ fq -d xing dv xing
 0x70|      ff                                       |  .             |    [98]: 255 entry 0x72-0x72.7 (1)
 0x70|         ff                                    |   .            |    [99]: 255 entry 0x73-0x73.7 (1)
 0x70|            00 00 00 00                        |    ....        |  quality: 0 0x74-0x77.7 (4)
-    |                                               |                |  lame_extension{}: 0x78-0x9b.7 (36)
-0x70|                        4c 61 76 63 35 38 2e 39|        Lavc58.9|    encoder: "Lavc58.91" 0x78-0x80.7 (9)
+0x70|                        4c 61 76 63 35 38 2e 39|        Lavc58.9|  encoder: "Lavc58.91" 0x78-0x80.7 (9)
 0x80|31                                             |1               |
-0x80|   00                                          | .              |    tag_revision: 0 0x81-0x81.3 (0.4)
-0x80|   00                                          | .              |    vbr_method: 0 0x81.4-0x81.7 (0.4)
-0x80|      00                                       |  .             |    lowpass_filter: 0 0x82-0x82.7 (1)
-0x80|         00 00 00 00                           |   ....         |    replay_gain_peak: 0 0x83-0x86.7 (4)
-0x80|                     00 00                     |       ..       |    radio_replay_gain: 0 0x87-0x88.7 (2)
-0x80|                           00 00               |         ..     |    audiophile_replay_gain: 0 0x89-0x8a.7 (2)
-0x80|                                 00            |           .    |    lame_flags: 0 0x8b-0x8b.3 (0.4)
-0x80|                                 00            |           .    |    lame_ath_type: 0 0x8b.4-0x8b.7 (0.4)
-0x80|                                    00         |            .   |    abr_vbr: 0 0x8c-0x8c.7 (1)
-0x80|                                       24 05   |             $. |    encoder_delay: 576 0x8d-0x8e.3 (1.4)
-0x80|                                          05 07|              ..|    encoder_padding: 1287 0x8e.4-0x8f.7 (1.4)
-0x90|00                                             |.               |    misc: 0 0x90-0x90.7 (1)
-0x90|   00                                          | .              |    mp3_gain: 0 0x91-0x91.7 (1)
-0x90|      00 00                                    |  ..            |    preset: 0 0x92-0x93.7 (2)
-0x90|            00 00 04 13                        |    ....        |    length: 1043 0x94-0x97.7 (4)
-0x90|                        c2 aa                  |        ..      |    music_crc: 49834 0x98-0x99.7 (2)
-0x90|                              92 0f|           |          ..|   |    tag_crc: 37391 0x9a-0x9b.7 (2)
+0x80|   00                                          | .              |  tag_revision: 0 0x81-0x81.3 (0.4)
+0x80|   00                                          | .              |  vbr_method: 0 0x81.4-0x81.7 (0.4)
+0x80|      00                                       |  .             |  lowpass_filter: 0 0x82-0x82.7 (1)
+0x80|         00 00 00 00                           |   ....         |  replay_gain_peak: 0 0x83-0x86.7 (4)
+0x80|                     00 00                     |       ..       |  radio_replay_gain: 0 0x87-0x88.7 (2)
+0x80|                           00 00               |         ..     |  audiophile_replay_gain: 0 0x89-0x8a.7 (2)
+0x80|                                 00            |           .    |  lame_flags: 0 0x8b-0x8b.3 (0.4)
+0x80|                                 00            |           .    |  lame_ath_type: 0 0x8b.4-0x8b.7 (0.4)
+0x80|                                    00         |            .   |  abr_vbr: 0 0x8c-0x8c.7 (1)
+0x80|                                       24 05   |             $. |  encoder_delay: 576 0x8d-0x8e.3 (1.4)
+0x80|                                          05 07|              ..|  encoder_padding: 1287 0x8e.4-0x8f.7 (1.4)
+0x90|00                                             |.               |  misc: 0 0x90-0x90.7 (1)
+0x90|   00                                          | .              |  mp3_gain: 0 0x91-0x91.7 (1)
+0x90|      00 00                                    |  ..            |  preset: 0 0x92-0x93.7 (2)
+0x90|            00 00 04 13                        |    ....        |  length: 1043 0x94-0x97.7 (4)
+0x90|                        c2 aa                  |        ..      |  music_crc: 49834 0x98-0x99.7 (2)
+0x90|                              92 0f|           |          ..|   |  tag_crc: 37391 0x9a-0x9b.7 (2)
