+++ pkg/interp/testdata/value_object.fqtest
@@ -172,30 +172,22 @@ mp3> .headers[0].flags.a = 1
         "scfsi0": 0,
         "share": 0
       },
-      "xing": {
+      "tags": {
+        "abr_vbr": 0,
+        "audiophile_replay_gain": 0,
         "bytes": 599,
+        "encoder": "Lavc58.91",
+        "encoder_delay": 576,
+        "encoder_padding": 1287,
         "frames": 2,
         "header": "Info",
-        "lame_extension": {
-          "abr_vbr": 0,
-          "audiophile_replay_gain": 0,
-          "encoder": "Lavc58.91",
-          "encoder_delay": 576,
-          "encoder_padding": 1287,
-          "lame_ath_type": 0,
-          "lame_flags": 0,
-          "length": 599,
-          "lowpass_filter": 0,
-          "misc": 0,
-          "mp3_gain": 0,
-          "music_crc": 25328,
-          "preset": 0,
-          "radio_replay_gain": 0,
-          "replay_gain_peak": 0,
-          "tag_crc": 23093,
-          "tag_revision": 0,
-          "vbr_method": 0
-        },
+        "lame_ath_type": 0,
+        "lame_flags": 0,
+        "length": 599,
+        "lowpass_filter": 0,
+        "misc": 0,
+        "mp3_gain": 0,
+        "music_crc": 25328,
         "present_flags": {
           "bytes": true,
           "frames": true,
@@ -203,7 +195,12 @@ mp3> .headers[0].flags.a = 1
           "toc": true,
           "unused": 0
         },
+        "preset": 0,
         "quality": 0,
+        "radio_replay_gain": 0,
+        "replay_gain_peak": 0,
+        "tag_crc": 23093,
+        "tag_revision": 0,
         "toc": [
           0,
           166,
@@ -305,7 +302,8 @@ mp3> .headers[0].flags.a = 1
           255,
           255,
           255
-        ]
+        ],
+        "vbr_method": 0
       }
     },
     {
@@ -541,30 +539,22 @@ mp3> .headers[0].flags.a |= empty
         "scfsi0": 0,
         "share": 0
       },
-      "xing": {
+      "tags": {
+        "abr_vbr": 0,
+        "audiophile_replay_gain": 0,
         "bytes": 599,
+        "encoder": "Lavc58.91",
+        "encoder_delay": 576,
+        "encoder_padding": 1287,
         "frames": 2,
         "header": "Info",
-        "lame_extension": {
-          "abr_vbr": 0,
-          "audiophile_replay_gain": 0,
-          "encoder": "Lavc58.91",
-          "encoder_delay": 576,
-          "encoder_padding": 1287,
-          "lame_ath_type": 0,
-          "lame_flags": 0,
-          "length": 599,
-          "lowpass_filter": 0,
-          "misc": 0,
-          "mp3_gain": 0,
-          "music_crc": 25328,
-          "preset": 0,
-          "radio_replay_gain": 0,
-          "replay_gain_peak": 0,
-          "tag_crc": 23093,
-          "tag_revision": 0,
-          "vbr_method": 0
-        },
+        "lame_ath_type": 0,
+        "lame_flags": 0,
+        "length": 599,
+        "lowpass_filter": 0,
+        "misc": 0,
+        "mp3_gain": 0,
+        "music_crc": 25328,
         "present_flags": {
           "bytes": true,
           "frames": true,
@@ -572,7 +562,12 @@ mp3> .headers[0].flags.a |= empty
           "toc": true,
           "unused": 0
         },
+        "preset": 0,
         "quality": 0,
+        "radio_replay_gain": 0,
+        "replay_gain_peak": 0,
+        "tag_crc": 23093,
+        "tag_revision": 0,
         "toc": [
           0,
           166,
@@ -674,7 +669,8 @@ mp3> .headers[0].flags.a |= empty
           255,
           255,
           255
-        ]
+        ],
+        "vbr_method": 0
       }
     },
     {
